refactor: split reliable tsa test
